Changed codebase to be foldered (breaking compiler's conventions)
I now generate the layout the compiler wants, eventually I'll just have a custom policy so that the compiler can accept the non-idiomatic layout

See scripts/build.ps1 & gen_staged_compiler_codebase.ps1 for how this is handled.

+++ scripts/clean.ps1
@@ -1,9 +1,9 @@
 cls
 
-$path_root  = git rev-parse --show-toplevel
-$path_code  = join-path $path_root 'code'
-$path_build = join-path $path_root 'build'
+$path_root                 = git rev-parse --show-toplevel
+$path_code                 = join-path $path_root 'code'
+$path_code_compiler_staged = Join-Path $path_root 'code_compiler_staged'
+$path_build                = join-path $path_root 'build'
 
-if ( test-path $path_build ) {
-	Remove-Item $path_build -Verbose -Force -Recurse
-}
+if ( test-path $path_build )               { Remove-Item $path_build -Verbose -Force -Recurse }
+if ( test-path $path_code_compiler_staged) { Remove-Item $path_code_compiler_staged -Verbose -Force -Recurse }
